Removes BaseAppRedirector:
* Disconnects AppRouter.
* Creates wrapping component to lookup + propagate the experiment.
* Moves GA pageview logic to App component.
* Moves AppRouter state processing to App.

Closes #1583.
